@@ -1,10 +1,10 @@
 [package]
 name = "samply-api"
-version = "0.23.0"
+version = "0.23.1"
 authors = ["Markus Stange <mstange.moz@gmail.com>"]
 license = "MIT OR Apache-2.0"
 edition = "2021"
-rust-version = "1.67"
+rust-version = "1.70" # needed by samply-symbols -> linux-perf-data -> prost-derive
 description = "JSON API for querying symbol information, uses samply-symbols."
 readme = "README.md"
 
@@ -13,18 +13,18 @@ default = []
 send_futures = ["samply-symbols/send_futures"]
 
 [dependencies]
-samply-symbols = { version = "0.22.0", path = "../samply-symbols" }
-thiserror = "1.0.61"
-serde = "1.0.202"
+samply-symbols = { version = "0.23.0", path = "../samply-symbols" }
+thiserror = "1.0.62"
+serde = "1.0.204"
 serde_derive = "1.0.188"
-serde_json = "1.0.117"
-yaxpeax-arch = { version = "0.2.8", default-features = false }
-yaxpeax-x86 = { version = "1.1.4", default-features = false, features = ["std", "fmt"] }
-yaxpeax-arm = { version = "0.2.5", default-features = false, features = ["std"] }
+serde_json = "1.0.120"
+yaxpeax-arch = { version = "0.3", default-features = false }
+yaxpeax-x86 = { version = "2", default-features = false, features = ["std", "fmt"] }
+yaxpeax-arm = { version = "0.3", default-features = false, features = ["std"] }
 
 [dev-dependencies]
 memmap2 = "0.9.4"
 anyhow = "1.0.86"
 futures = "0.3.5"
 assert-json-diff = "2.0.1"
-serde_json = "1.0.117"
+serde_json = "1.0.120"
