@@ -3,7 +3,7 @@ name = "usamply"
 version = "0.12.8"
 authors = ["Markus Stange <mstange@themasta.com>", "Vladimir Vukicevic <vladimir@pobox.com>"]
 edition = "2021"
-rust-version = "1.74" # needed by clap and wholesym
+rust-version = "1.75" # needed by wholesym -> fs4
 license = "MIT OR Apache-2.0"
 description = "A command line profiler for macOS and Linux. (usamply customized from samply)"
 repository = "https://github.com/Unity-Technologies/usamply/"
@@ -13,14 +13,14 @@ readme = "README.md"
 
 fxprof-processed-profile = { version = "0.7", path = "../fxprof-processed-profile" }
 # framehop = { path = "../../framehop" }
-framehop = "0.11.2"
+framehop = "0.12"
 # linux-perf-data = { path = "../../linux-perf-data" }
 linux-perf-data = "0.10.1"
 
-tokio = { version = "1.37.0", features = ["rt", "rt-multi-thread", "macros"] }
+tokio = { version = "1.38.0", features = ["rt", "rt-multi-thread", "macros"] }
 tokio-util = "0.7.11"
 hyper = { version = "1", features = ["full"] }
-hyper-util = { version = "0.1.3", features = ["server", "http1", "tokio"] }
+hyper-util = { version = "0.1.6", features = ["server", "http1", "tokio"] }
 http-body-util = "0.1"
 futures-util = "0.3"
 clap = { version = "4", features = ["derive"] }
@@ -28,10 +28,10 @@ byteorder = "1.4.3"
 debugid = "0.8.0"
 memchr = "2.7.2"
 memmap2 = "0.9.4"
-serde_json = "1.0.117"
-thiserror = "1.0.61"
+serde_json = "1.0.120"
+thiserror = "1.0.62"
 tempfile = "3.10.1"
-uuid = {  version = "1.8.0", features = ["v4"] }
+uuid = {  version = "1.10.0", features = ["v4"] }
 percent-encoding = "2.1.0"
 libc = "0.2.155"
 flate2 = "1.0"
@@ -39,9 +39,9 @@ opener = { version = "0.7.1", default-features = false }
 rand = "0.8.4"
 nix-base32 = "0.1.1"
 serde_derive = "1.0.137"
-serde = "1.0.202"
-wholesym = { version = "0.5.0", path = "../wholesym", features = ["api"]}
-dirs = "5.0.0"
+serde = "1.0.204"
+wholesym = { version = "0.7.0", path = "../wholesym", features = ["api"]}
+platform-dirs = "0.3"
 once_cell = "1.17"
 fxhash = "0.2.1"
 mio = { version = "0.8.11", features = ["os-ext", "os-poll"] }
@@ -50,6 +50,8 @@ log = "0.4.21"
 env_logger = "0.11.3"
 cfg-if = "1.0.0"
 fs4 = "0.8.3"
+humantime = "2.1.0"
+shlex = "1.3.0"
 
 [target.'cfg(any(target_os = "android", target_os = "macos", target_os = "linux"))'.dependencies]
 
@@ -59,17 +61,19 @@ crossbeam-channel = "0.5.13"
 [target.'cfg(target_os = "macos")'.dependencies]
 
 mach = "0.3.2"
-lazy_static = "1.4.0"
+lazy_static = "1.5.0"
 flate2 = "1.0.30"
 sysctl = "0.5.4"
 tempfile = "3.10.1"
+plist = "1.7.0"
 
 [target.'cfg(any(target_os = "android", target_os = "linux"))'.dependencies]
 
 parking_lot = "0.12.2"
 num_cpus = "1.13.1"
 uname = "0.1.1"
-nix = { version = "0.28", features = ["fs", "process"] }
+nix = { version = "0.29", features = ["fs", "process"] }
+os-release = "0.1.0"
 
 [target.'cfg(windows)'.dependencies]
 
@@ -80,11 +84,12 @@ num-traits = "0.2"
 num-derive = "0.4"
 runas = "1.2.0"
 which = "6.0.1"
+winver = "1"
 etw-reader = { path = "../etw-reader" }
 # linux-perf-data = "0.10.1"
 
 [target.'cfg(windows)'.dependencies.windows]
-version = "0.56"
+version = "0.58"
 features =  ["Win32",
              "Win32_Foundation",
              "Win32_Security",
@@ -105,4 +110,4 @@ features =  ["Win32",
 [dependencies.object]
 default-features = false
 features = ["std", "read_core", "elf", "pe", "unaligned", "write"]
-version = "0.35"
+version = "0.36"
