@@ -1,9 +1,9 @@
 [package]
 name = "wholesym"
-version = "0.5.0"
+version = "0.7.0"
 authors = ["Markus Stange <mstange.moz@gmail.com>"]
 edition = "2021"
-rust-version = "1.74" # needed by symsrv -> cab -> lxzd (which runs into unjustified "private type in public interface" errors with 1.73 and lower)
+rust-version = "1.75" # needed by fs4
 description = "A complete solution for fetching symbol files and resolving code addresses to symbols and debuginfo."
 readme = "README.md"
 homepage = "https://github.com/mstange/samply/tree/main/wholesym"
@@ -17,24 +17,24 @@ api = ["samply-api"]
 
 [dependencies]
 debugid = "0.8.0"
-samply-api = { version = "0.23.0", path = "../samply-api", features = ["send_futures"], optional = true }
-samply-symbols = { version = "0.22.0", path = "../samply-symbols", features = ["send_futures"] }
+samply-api = { version = "0.23.1", path = "../samply-api", features = ["send_futures"], optional = true }
+samply-symbols = { version = "0.23.0", path = "../samply-symbols", features = ["send_futures"] }
 # symsrv = { path = "../../symsrv" }
 symsrv = "0.5.1"
 yoke = "0.7"
 yoke-derive = "0.7"
 libc = "0.2.155"
 uuid = "1"
-reqwest = { version = "0.12.4", default-features = false, features = [
+reqwest = { version = "0.12", default-features = false, features = [
     "rustls-tls",
     "stream",
     "gzip",
     "brotli",
     "deflate"
 ] }
-bytes = "1.1.0"
+bytes = "1.6.1"
 memmap2 = "0.9.4"
-tokio = { version = "1.17.0", features = ["fs"] }
+tokio = { version = "1.38.0", features = ["fs"] }
 futures-util = "0.3.30"
 
 # Needed for moria_mac_spotlight, to find dSYM files
@@ -44,4 +44,4 @@ core-foundation = "0.9.1"
 
 [dev-dependencies]
 futures = "0.3.5"
-tokio = { version = "1.17.0", features = ["macros"] } # Feature "macros" for #[tokio::test]
+tokio = { version = "1.38.0", features = ["macros"] } # Feature "macros" for #[tokio::test]
