@@ -1,6 +1,5 @@
 use std::collections::HashMap;
 use std::fs::File;
-use std::io::BufWriter;
 use std::ops::Deref;
 use std::os::unix::process::ExitStatusExt;
 use std::path::Path;
@@ -29,6 +28,8 @@ use crate::shared::ctrl_c::CtrlC;
 use crate::shared::recording_props::{
     ProcessLaunchProps, ProfileCreationProps, RecordingMode, RecordingProps,
 };
+use crate::shared::save_profile::save_profile_to_file;
+use crate::shared::symbol_props::SymbolProps;
 
 #[cfg(target_arch = "x86_64")]
 pub type ConvertRegsNative = crate::linux_shared::ConvertRegsX86_64;
@@ -40,6 +41,7 @@ pub fn start_recording(
     recording_mode: RecordingMode,
     recording_props: RecordingProps,
     profile_creation_props: ProfileCreationProps,
+    symbol_props: SymbolProps,
     server_props: Option<ServerProps>,
 ) -> Result<ExitStatus, ()> {
     let process_launch_props = match recording_mode {
@@ -50,7 +52,13 @@ pub fn start_recording(
             std::process::exit(1)
         }
         RecordingMode::Pid(pid) => {
-            start_profiling_pid(pid, recording_props, profile_creation_props, server_props);
+            start_profiling_pid(
+                pid,
+                recording_props,
+                profile_creation_props,
+                symbol_props,
+                server_props,
+            );
             return Ok(ExitStatus::from_raw(0));
         }
         RecordingMode::Launch(process_launch_props) => process_launch_props,
@@ -96,6 +104,15 @@ pub fn start_recording(
     let output_file_copy = recording_props.output_file.clone();
     let interval = recording_props.interval;
     let time_limit = recording_props.time_limit;
+    let initial_exec_name = command_name.to_string_lossy().to_string();
+    let initial_cmdline: Vec<String> = std::iter::once(initial_exec_name.clone())
+        .chain(args.iter().map(|arg| arg.to_string_lossy().to_string()))
+        .collect();
+    let initial_exec_name = match initial_exec_name.rfind('/') {
+        Some(pos) => initial_exec_name[pos + 1..].to_string(),
+        None => initial_exec_name,
+    };
+    let initial_exec_name_and_cmdline = (initial_exec_name, initial_cmdline);
     let observer_thread = thread::spawn(move || {
         let unstable_presymbolicate = profile_creation_props.unstable_presymbolicate;
         let mut converter = make_converter(interval, profile_creation_props);
@@ -130,6 +147,7 @@ pub fn start_recording(
             profile_another_pid_reply_sender,
             stop_receiver,
             unstable_presymbolicate,
+            Some(initial_exec_name_and_cmdline),
         );
     });
 
@@ -236,7 +254,7 @@ pub fn start_recording(
         )
         .expect("Couldn't parse libinfo map from profile file");
 
-        start_server_main(profile_filename, server_props, libinfo_map);
+        start_server_main(profile_filename, server_props, symbol_props, libinfo_map);
     }
 
     let exit_status = match wait_status {
@@ -250,6 +268,7 @@ fn start_profiling_pid(
     pid: u32,
     recording_props: RecordingProps,
     profile_creation_props: ProfileCreationProps,
+    symbol_props: SymbolProps,
     server_props: Option<ServerProps>,
 ) {
     // When the first Ctrl+C is received, stop recording.
@@ -289,6 +308,7 @@ fn start_profiling_pid(
                 profile_another_pid_reply_sender,
                 ctrl_c_receiver,
                 unstable_presymbolicate,
+                None,
             )
         }
     });
@@ -328,7 +348,7 @@ fn start_profiling_pid(
         )
         .expect("Couldn't parse libinfo map from profile file");
 
-        start_server_main(&output_file, server_props, libinfo_map);
+        start_server_main(&output_file, server_props, symbol_props, libinfo_map);
     }
 }
 
@@ -366,19 +386,27 @@ fn make_converter(
         event_names: vec!["cycles".to_string()],
     };
 
-    Converter::<framehop::UnwinderNative<MmapRangeOrVec, framehop::MayAllocateDuringUnwind>>::new(
+    let mut converter = Converter::<
+        framehop::UnwinderNative<MmapRangeOrVec, framehop::MayAllocateDuringUnwind>,
+    >::new(
         &profile_creation_props,
         ReferenceTimestamp::from_system_time(SystemTime::now()),
-        None,
+        profile_creation_props.profile_name(),
         HashMap::new(),
         machine_info.as_ref().map(|info| info.release.as_str()),
         first_sample_time,
         endian,
         framehop::CacheNative::new(),
-        None,
+        Vec::new(),
+        Vec::new(),
         interpretation,
         None,
-    )
+        false,
+    );
+    if let Ok(os_release) = os_release::OsRelease::new() {
+        converter.set_os_name(&os_release.pretty_name);
+    }
+    converter
 }
 
 fn init_profiler(
@@ -454,12 +482,20 @@ fn init_profiler(
         }
     };
 
+    let (exe_name, cmdline) = get_process_cmdline(pid).expect("Couldn't read process cmdline");
+    let comm_data = std::fs::read(format!("/proc/{pid}/comm")).expect("Couldn't read process comm");
+    let length = memchr::memchr(b'\0', &comm_data).unwrap_or(comm_data.len());
+    let comm_name = std::str::from_utf8(&comm_data[..length])
+        .unwrap()
+        .trim_end();
+    converter.register_existing_process(pid as i32, comm_name, &exe_name, cmdline);
+
     // TODO: Gather threads / processes recursively, here and in PerfGroup setup.
-    for entry in std::fs::read_dir(format!("/proc/{pid}/task"))
+    for thread_entry in std::fs::read_dir(format!("/proc/{pid}/task"))
         .unwrap()
         .flatten()
     {
-        let tid: u32 = entry.file_name().to_string_lossy().parse().unwrap();
+        let tid: u32 = thread_entry.file_name().to_string_lossy().parse().unwrap();
         let comm_path = format!("/proc/{pid}/task/{tid}/comm");
         if let Ok(buffer) = std::fs::read(comm_path) {
             let length = memchr::memchr(b'\0', &buffer).unwrap_or(buffer.len());
@@ -548,6 +584,7 @@ fn run_profiler(
     more_processes_reply_sender: Sender<bool>,
     mut stop_receiver: oneshot::Receiver<()>,
     unstable_presymbolicate: bool,
+    mut initial_exec_name_and_cmdline: Option<(String, Vec<String>)>,
 ) {
     // eprintln!("Running...");
 
@@ -632,7 +669,27 @@ fn run_profiler(
                     converter.handle_fork(e);
                 }
                 EventRecord::Comm(e) => {
-                    converter.handle_comm(e, record.timestamp());
+                    if e.is_execve {
+                        // Try to get the command line arguments for this process.
+                        let exec_name_and_cmdline =
+                            if let Some(initial) = initial_exec_name_and_cmdline.take() {
+                                // This COMM event is the first exec that we're processing. If we get
+                                // here, it means we're in the "launch process" case and we're seeing
+                                // the exec for that initial launched process.
+                                Some(initial)
+                            } else {
+                                // Attempt to get the process cmdline from /proc/{pid}/cmdline.
+                                // This isn't very reliable because we're processing the perf event records
+                                // in batches, with a delay, so the COMM record may be old enough that the
+                                // pid no longer exists, or the pid may even refer to a different process now.
+                                // Unfortunately there are no perf event records that give us the process
+                                // command line.
+                                get_process_cmdline(e.pid as u32).ok()
+                            };
+                        converter.handle_exec(e, record.timestamp(), exec_name_and_cmdline);
+                    } else {
+                        converter.handle_thread_rename(e, record.timestamp());
+                    }
                 }
                 EventRecord::Exit(e) => {
                     converter.handle_exit(e);
@@ -673,11 +730,7 @@ fn run_profiler(
 
     let profile = converter.finish();
 
-    {
-        let output_file = File::create(output_filename).unwrap();
-        let writer = BufWriter::new(output_file);
-        serde_json::to_writer(writer, &profile).expect("Couldn't write JSON");
-    }
+    save_profile_to_file(&profile, output_filename).expect("Couldn't write JSON");
 
     if unstable_presymbolicate {
         crate::shared::symbol_precog::presymbolicate(
@@ -690,4 +743,23 @@ fn run_profiler(
 pub fn read_string_lossy<P: AsRef<Path>>(path: P) -> std::io::Result<String> {
     let data = std::fs::read(path)?;
     Ok(String::from_utf8_lossy(&data).into_owned())
+}
+
+fn get_process_cmdline(pid: u32) -> std::io::Result<(String, Vec<String>)> {
+    let cmdline_bytes = std::fs::read(format!("/proc/{pid}/cmdline"))?;
+    let mut remaining_bytes = &cmdline_bytes[..];
+    let mut cmdline = Vec::new();
+    while let Some(nul_byte_pos) = memchr::memchr(b'\0', remaining_bytes) {
+        let arg_slice = &remaining_bytes[..nul_byte_pos];
+        remaining_bytes = &remaining_bytes[nul_byte_pos + 1..];
+        cmdline.push(String::from_utf8_lossy(arg_slice).to_string());
+    }
+
+    let exe_arg = &cmdline[0];
+    let exe_name = match exe_arg.rfind('/') {
+        Some(pos) => exe_arg[pos + 1..].to_string(),
+        None => exe_arg.to_string(),
+    };
+
+    Ok((exe_name, cmdline))
 }
