@@ -1,18 +1,17 @@
-use std::fs::File;
-use std::io::BufWriter;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 
 use fxprof_processed_profile::{Profile, ReferenceTimestamp, SamplingInterval};
-use serde_json::to_writer;
 
 use super::etw_gecko;
 use crate::shared::included_processes::IncludedProcesses;
 use crate::shared::recording_props::ProfileCreationProps;
+use crate::shared::save_profile::save_profile_to_file;
 use crate::windows::profile_context::ProfileContext;
 
 pub fn convert_etl_file_to_profile(
     filename: &Path,
-    output_filename: &Path,
+    extra_etl_filenames: &[PathBuf],
+    output_file: &Path,
     profile_creation_props: ProfileCreationProps,
     included_processes: Option<IncludedProcesses>,
 ) {
@@ -21,26 +20,22 @@ pub fn convert_etl_file_to_profile(
 
     let interval_8khz = SamplingInterval::from_nanos(122100); // 8192Hz // only with the higher recording rate?
     let profile = Profile::new(
-        &profile_creation_props.profile_name,
+        profile_creation_props.profile_name(),
         timebase,
         interval_8khz, // recording_props.interval.into(),
     );
 
     let arch = get_native_arch(); // TODO: Detect arch from file
 
+    eprintln!("Processing ETL trace...");
+
     let mut context =
         ProfileContext::new(profile, arch, included_processes, profile_creation_props);
 
-    etw_gecko::profile_pid_from_etl_file(&mut context, filename);
+    etw_gecko::process_etl_files(&mut context, filename, extra_etl_filenames);
 
     let profile = context.finish();
-
-    // write the profile to a json file
-    let file = File::create(output_filename).unwrap();
-    let writer = BufWriter::new(file);
-    {
-        to_writer(writer, &profile).expect("Couldn't write JSON");
-    }
+    save_profile_to_file(&profile, output_file).expect("Couldn't write JSON");
 }
 
 #[cfg(target_arch = "x86")]
